umemployment-rate-analysis/requirements.txt
selenium==3.141.0
beautifulsoup4==4.9.2
PyYAML==5.3.1
